Code Improvements on API

+++ moldme/Controllers/ChatController.cs
@@ -1,12 +1,14 @@
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.IdentityModel.Tokens;
 using moldme.data;
+using moldme.Interface;
 using moldme.Models;
 
-namespace moldme.Controllers
-{
+namespace moldme.Controllers;
+
     [ApiController]
     [Route("api/[controller]")]
-    public class ChatController : ControllerBase
+    public class ChatController : ControllerBase, IChat
     {
         private readonly ApplicationDbContext _context;
 
@@ -15,31 +17,42 @@ namespace moldme.Controllers
             _context = context;
         }
 
-        // Create
-        [HttpPost]
-        public async Task<ActionResult<Chat>> CreateChat(Chat chat)
+        ///<inheritdoc cref="IChat.ChatCreate(string)"/>
+        [HttpPost("createChat")]
+        public async Task<ActionResult<Chat>> ChatCreate(String ProjectId)
         {
+            if (ProjectId.IsNullOrEmpty())
+                return BadRequest("Project Id is required");
+
+            Project Project = _context.Projects.Find(ProjectId);
+
+            if (Project == null)
+                return NotFound("Project not found");
+
+            Chat chat = new Chat();
+            chat.ProjectId = ProjectId;
+            chat.Project = Project;
+
             _context.Chats.Add(chat);
             _context.SaveChanges();
 
             return Ok("Chat created successfully");
         }
 
-        // Delete
-        [HttpDelete("{id}")]
-        public IActionResult Delete(String id)
+        ///<inheritdoc cref="IChat.ChatDelete(string)"/>
+        [HttpDelete("deleteChat/{ChatId}")]
+        public IActionResult ChatDelete(String ChatId)
         {
-            var chat = _context.Chats.Find(id);
-            if (chat == null)
-            {
-                return NotFound();
-            }
+            if (ChatId.IsNullOrEmpty())
+                return BadRequest("Chat Id is required");
 
+            var chat = _context.Chats.Find(ChatId);
+            if (chat == null)
+                return NotFound();
+            
             _context.Chats.Remove(chat);
             _context.SaveChanges();
 
             return Ok("Chat deleted successfully");
         }
     }
-    
-}
